requirements.txt
streamlit
PyYAML
bcrypt